backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
python-dotenv
deepface
opencv-python
groq
pillow